requirements.txt
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework_simplejwt==4.8.0
dj-database-url==0.5.0
drf-yasg==1.20.0
flake8==3.8.3
gunicorn==20.0.4
model-mommy==2.0.0
psycopg2-binary==2.8.6
whitenoise==5.2.0
